Delete classes of old job system and refactor
