+++ src/main/scala/com/amazon/deequ/backend/jobmanagement/JobManagement.scala
@@ -2,34 +2,11 @@ package com.amazon.deequ.backend.jobmanagement
 
 import java.util.UUID.randomUUID
 
-import com.amazon.deequ.backend.jobmanagement.analyzerJobs._
 import com.amazon.deequ.backend.jobmanagement.extractors.{DistinctnessAnalyzerExtractor, UniquenessAnalyzerExtractor}
 import org.json4s.JValue
 
-import scala.collection.immutable.ListMap
-
 class JobManagement {
   private var jobs = synchronized(Map[String, ExecutableAnalyzerJob]())
-
-  private val availableAnalyzers = ListMap[String, AnalyzerJob[_]](
-    "completeness" -> CompletenessAnalyzerJob,
-    "compliance" -> ComplianceAnalyzerJob,
-    "correlation" -> CorrelationAnalyzerJob,
-    "countDistinct" -> CountDistinctAnalyzerJob,
-    "dataType" -> DataTypeAnalyzerJob,
-    "distinctness" -> DistinctnessAnalyzerJob,
-    "entropy" -> EntropyAnalyzerJob,
-    "histogram" -> HistogramAnalyzerJob,
-    "maximum" -> MaximumAnalyzerJob,
-    "mean" -> MeanAnalyzerJob,
-    "minimum" -> MinimumAnalyzerJob,
-    "patternMatch" -> PatternMatchAnalyzerJob,
-    "size" -> SizeAnalyzerJob,
-    "standardDeviation" -> StandardDeviationAnalyzerJob,
-    "sum" -> SumAnalyzerJob,
-    "uniqueness" -> UniquenessAnalyzerJob,
-    "uniqueValueRatio" -> UniqueValueRatioAnalyzerJob
-  )
 
   def getAvailableAnalyzers: Seq[Map[String, Any]] = {
     AnalysisRun.availableExtractors.map {
@@ -103,24 +80,6 @@ class JobManagement {
         }
         m
     }.toSeq
-  }
-
-  def startJob(requestedAnalyzer: String, params: JValue): String = {
-    val jobId = randomUUID().toString.replace("-", "")
-
-    if (!availableAnalyzers.exists(_._1 == requestedAnalyzer)) {
-      throw new NoSuchAnalyzerException(
-        s"There is no analyzer called $requestedAnalyzer. " +
-          s"Available analyzers are ${availableAnalyzers.keys.mkString("[", ", ", "]")}")
-    }
-
-    val analyzer = availableAnalyzers(requestedAnalyzer)
-
-    val job = analyzer.from(params)
-    jobs += (jobId -> job)
-    job.start()
-
-    jobId
   }
 
   def startJobs(tableName: String, context: String, parsedBody: JValue): Any = {
