--- src/main/scala/com/amazon/deequ/backend/jobmanagement/analyzerJobs/MeanAnalyzerJob.scala
@@ -1,30 +1,0 @@
-package com.amazon.deequ.backend.jobmanagement.analyzerJobs
-
-import com.amazon.deequ.analyzers.jdbc._
-import com.amazon.deequ.analyzers.{Mean, MeanState}
-import com.amazon.deequ.backend.jobmanagement.{AnalyzerJob, ColumnAndWhereAnalyzerParams, RequestParameter}
-import com.amazon.deequ.metrics.DoubleMetric
-import org.json4s.JValue
-
-object MeanAnalyzerJob extends AnalyzerJob[ColumnAndWhereAnalyzerParams] {
-
-  val name = "Mean"
-  val description = " This is the average over all values of the given column. Only for numeric columns."
-
-  val acceptedRequestParams: () => Array[RequestParameter] =
-    () => extractFieldNames[ColumnAndWhereAnalyzerParams]
-
-  def extractFromJson(requestParams: JValue): ColumnAndWhereAnalyzerParams = {
-    requestParams.extract[ColumnAndWhereAnalyzerParams]
-  }
-
-  def funcWithJdbc(params: ColumnAndWhereAnalyzerParams): Any = {
-    analyzerWithJdbc[MeanState, DoubleMetric, JdbcMean](
-      JdbcMean(params.column, params.where), params.table)
-  }
-
-  def funcWithSpark(params: ColumnAndWhereAnalyzerParams): Any = {
-    analyzerWithSpark[MeanState, DoubleMetric, Mean](
-      Mean(params.column, params.where), params.table)
-  }
-}
