--- src/main/scala/com/amazon/deequ/backend/jobmanagement/analyzerJobs/SizeAnalyzerJob.scala
@@ -1,30 +1,0 @@
-package com.amazon.deequ.backend.jobmanagement.analyzerJobs
-
-import com.amazon.deequ.analyzers.jdbc._
-import com.amazon.deequ.analyzers.{NumMatches, Size}
-import com.amazon.deequ.backend.jobmanagement.{AnalyzerJob, ColumnAndWhereAnalyzerParams, RequestParameter}
-import com.amazon.deequ.metrics.DoubleMetric
-import org.json4s.JValue
-
-object SizeAnalyzerJob extends AnalyzerJob[ColumnAndWhereAnalyzerParams] {
-
-  val name = "Size"
-  val description = "Is the amount of values in the given column."
-
-  val acceptedRequestParams: () => Array[RequestParameter] =
-    () => extractFieldNames[ColumnAndWhereAnalyzerParams]
-
-  def extractFromJson(requestParams: JValue): ColumnAndWhereAnalyzerParams = {
-    requestParams.extract[ColumnAndWhereAnalyzerParams]
-  }
-
-  def funcWithJdbc(params: ColumnAndWhereAnalyzerParams): Any = {
-    analyzerWithJdbc[NumMatches, DoubleMetric, JdbcSize](
-      JdbcSize(params.where), params.table)
-  }
-
-  def funcWithSpark(params: ColumnAndWhereAnalyzerParams): Any = {
-    analyzerWithSpark[NumMatches, DoubleMetric, Size](
-      Size(params.where), params.table)
-  }
-}
