--- src/main/scala/com/amazon/deequ/backend/jobmanagement/analyzerJobs/SumAnalyzerJob.scala
@@ -1,30 +1,0 @@
-package com.amazon.deequ.backend.jobmanagement.analyzerJobs
-
-import com.amazon.deequ.analyzers.jdbc._
-import com.amazon.deequ.analyzers.{Sum, SumState}
-import com.amazon.deequ.backend.jobmanagement.{AnalyzerJob, ColumnAndWhereAnalyzerParams, RequestParameter}
-import com.amazon.deequ.metrics.DoubleMetric
-import org.json4s.JValue
-
-object SumAnalyzerJob extends AnalyzerJob[ColumnAndWhereAnalyzerParams] {
-
-  val name = "Sum"
-  val description = "The sum of all values in the given column. Only for numeric columns."
-
-  val acceptedRequestParams: () => Array[RequestParameter] =
-    () => extractFieldNames[ColumnAndWhereAnalyzerParams]
-
-  def extractFromJson(requestParams: JValue): ColumnAndWhereAnalyzerParams = {
-    requestParams.extract[ColumnAndWhereAnalyzerParams]
-  }
-
-  def funcWithJdbc(params: ColumnAndWhereAnalyzerParams): Any = {
-    analyzerWithJdbc[SumState, DoubleMetric, JdbcSum](
-      JdbcSum(params.column, params.where), params.table)
-  }
-
-  def funcWithSpark(params: ColumnAndWhereAnalyzerParams): Any = {
-    analyzerWithSpark[SumState, DoubleMetric, Sum](
-      Sum(params.column, params.where), params.table)
-  }
-}
