+++ src/main/scala/com/amazon/deequ/backend/jobmanagement/extractors/CorrelationAnalyzerExtractor.scala
@@ -2,9 +2,14 @@ package com.amazon.deequ.backend.jobmanagement.extractors
 
 import com.amazon.deequ.analyzers.Correlation
 import com.amazon.deequ.analyzers.jdbc.JdbcCorrelation
-import com.amazon.deequ.backend.jobmanagement.RequestParameter
-import com.amazon.deequ.backend.jobmanagement.analyzerJobs.CorrelationAnalyzerParams
+import com.amazon.deequ.backend.jobmanagement.{AnalyzerParams, RequestParameter}
 import org.json4s.JValue
+
+case class CorrelationAnalyzerParams(analyzer: String,
+                                     firstColumn: String,
+                                     secondColumn: String,
+                                     where: Option[String] = None)
+  extends AnalyzerParams
 
 object CorrelationAnalyzerExtractor extends AnalyzerExtractor[CorrelationAnalyzerParams] {
   override var params: CorrelationAnalyzerParams = _
